demo_project/requirements.txt
dal-admin-filters
Django>=2.1.7
django-autocomplete-light>=3.3.2
pytz
